# shared/api/openapi.yaml
openapi: 3.1.0
info:
  title: 112 AI Trainer REST API
  version: 0.3.0
  description: Backend scenario API and Voice Gateway API on separate ports.
servers:
- url: http://127.0.0.1:8000/api/v1
  description: Scenario backend
- url: http://127.0.0.1:8001/api/v1
  description: Voice Gateway
security:
- bearerAuth: []
paths:
...
  /instructor/dds/profiles:
    get:
      summary: Profiles
      operationId: profiles_api_v1_instructor_dds_profiles_get
      parameters:
      - name: X-User-Session
        in: header
        required: false
        schema:
          type: string
          default: ''
          title: X-User-Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
  /instructor/dds/incoming:
    get:
      summary: Incoming
      operationId: incoming_api_v1_instructor_dds_incoming_get
      parameters:
      - name: profile
        in: query
        required: true
        schema:
          type: string
          title: Profile
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 100
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: X-User-Session
        in: header
        required: false
        schema:
          type: string
          default: ''
          title: X-User-Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
...
  /instructor/dds/profiles/{profile}/recipients:
    get:
      summary: Recipient Settings
      operationId: recipient_settings_api_v1_instructor_dds_profiles__profile__recipients_get
      parameters:
      - name: profile
        in: path
        required: true
        schema:
          type: string
          pattern: ^(general|fire|police|medical|gas|utilities)$
          title: Profile
      - name: X-User-Session
        in: header
        required: false
        schema:
          type: string
          default: ''
          title: X-User-Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
    put:
      summary: Save Recipient Settings
      operationId: save_recipient_settings_api_v1_instructor_dds_profiles__profile__recipients_put
      parameters:
      - name: profile
        in: path
        required: true
        schema:
          type: string
          pattern: ^(general|fire|police|medical|gas|utilities)$
          title: Profile
      - name: X-User-Session
        in: header
        required: false
        schema:
          type: string
          default: ''
          title: X-User-Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipientSettings'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    RecipientSettings:
      properties:
        revision:
          type: integer
          minimum: 0.0
          title: Revision
        recipients:
          items:
            type: string
            maxLength: 160
            minLength: 1
          type: array
          maxItems: 20
          title: Recipients
      additionalProperties: false
      type: object
      required:
      - revision
      title: RecipientSettings
...